Windows:  Fix and suppress some analysis warnings
This fixes warnings that are emitted when building with the `-analyze`
flag enabled.

+++ modules/juce_dsp/frequency/juce_FFT_test.cpp
@@ -148,7 +148,7 @@ struct FFTUnitTest  : public UnitTest
                 HeapBlock<Complex<float>> frequency (n);
 
                 fillRandom (random, inout.getData(), n);
-                zeromem (reference.getData(), sizeof (float) * (n << 1));
+                zeromem (reference.getData(), sizeof (float) * ((size_t) n << 1));
                 performReferenceFourier (inout.getData(), frequency.getData(), n, false);
 
                 for (size_t i = 0; i < n; ++i)
